Add support for property names with dots
Add support for property names that are not valid MVEL identifiers
(e.g., a property name with a dot in it such as "foo.env") by adding
support for an "mvel(<properties-map-identifier>)" syntax for the
special "mvel" property name allowing a properties map identifier to
be specified.  The previous "mvel" syntax (i.e., the four-character
literal) works as before.

For example, if the special property name were

  mvel(p)

the value of the property "foo.env" could be accessed in the MVEL
expression with

  p["foo.env"]